"Upload scale-padding-fix.js"
Adds a JavaScript project called scale-padding-fix.js that reads the file "scale.txt" and outputs the data in a format that is easier to parse. The script first reads the contents of the file "scale.txt" and stores them in an array of strings. Then, it calculates the length of the longest number sequence in the file. Finally, it iterates over the lines in the file and pads the number sequences with zeros so that they are all the same length. The padded number sequences are then added to the output.