Initial commit of the scale-padding-fix project.
Initial commit of the scale-padding-fix project.

Description
This commit contains the initial version of the scale-padding-fix project. The project consists of the following files:

README.md
back-office-output.txt
scale.txt
scale-padding-fix.exe
The project pads number sequences in a file with zeros so that they are all the same length. The script first reads the contents of the file scale.txt and stores them in an array of strings. Then, it calculates the length of the longest number sequence in the file. Finally, it iterates over the lines in the file and pads the number sequences with zeros so that they are all the same length. The padded number sequences are then added to the output.

The project has been tested on a sample file and the output is correct.

Changes
Added the files README.md, back-office-output.txt, scale.txt, and scale-padding-fix.exe to the repository.